updated gulp and refactored main.js
