+++ .gitignore
@@ -1,3 +1,5 @@
 *.pyc
 node_modules
 .ctrlpignore
+frontend/static/js/all.js
+frontend/dist/*
