+++ frontend/index.html
@@ -4,7 +4,7 @@
     <title>Berkeley Power Rankings</title>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css"></link>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="frontend/static/css/style.css">
+    <link rel="stylesheet" href="frontend/dist/style.min.css">
   </head>
   <body>
     <div id="nav"></div>
@@ -21,6 +21,7 @@
     <script type="text/javascript" src="frontend/static/js/lib/fbuser.min.js"></script>
     <script type="text/javascript" src="frontend/static/js/lib/fuse.min.js"></script>
     <script type="text/javascript" src="frontend/static/js/all.js"></script>
+    <script type="text/javascript" src="frontend/dist/all.min.js"></script>
 
   </body>
 </html>
