+++ frontend/package.json
@@ -4,11 +4,14 @@
   "description": "",
   "main": "gulpfile.js",
   "devDependencies": {
-    "gulp": "^3.8.11",
     "browser-sync": "^2.3.1",
+    "gulp": "^3.8.11",
+    "gulp-concat": "^2.6.0",
+    "gulp-fixmyjs": "^1.0.2",
+    "gulp-minify-css": "^1.2.3",
     "gulp-plumber": "^1.0.0",
-    "gulp-uglify": "^1.1.0",
-    "gulp-concat": "^2.6.0"
+    "gulp-rename": "^1.2.2",
+    "gulp-uglify": "^1.1.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
