--- frontend/static/images/.desktop
@@ -1,6 +1,0 @@
-[Desktop Entry]
-Encoding=UTF-8
-Name=Link to 
-Type=Link
-URL=about:blank
-Icon=text-html
